+++ src/lib.rs
@@ -110,7 +110,8 @@
 pub use errors::LasZipError;
 #[cfg(feature = "parallel")]
 pub use laszip::parallel::{
-    par_compress_buffer, par_decompress_buffer, ParLasZipCompressor, ParLasZipDecompressor,
+    par_compress_buffer, par_decompress, par_decompress_buffer, ParLasZipCompressor,
+    ParLasZipDecompressor,
 };
 pub use laszip::{compress_buffer, decompress_buffer};
 pub use laszip::{